Run cleanup on push as part of main workflow
If not, there will be a race between the build job in main, and the
cleanup job in cleanup, resulting in incorrect or failing deletions.

Keep scheduled and dispatch in separate workflow for cleanup outside of pushes, using same concurrency group to avoid running concurrently with cleanup job in main.